+++ TabletopTweaks/Config/ModSettings.cs
@@ -8,11 +8,13 @@ namespace TabletopTweaks.Config {
         public static ModEntry ModEntry;
         public static Fixes Fixes;
         public static AddedContent AddedContent;
+        public static Homebrew Homebrew;
         public static Blueprints Blueprints;
 
         public static void LoadAllSettings() {
             LoadSettings("Fixes.json", ref Fixes);
             LoadSettings("AddedContent.json", ref AddedContent);
+            LoadSettings("Homebrew.json", ref Homebrew);
             LoadSettings("Blueprints.json", ref Blueprints);
         }
         private static void LoadSettings<T>(string fileName, ref T setting) where T : IUpdatableSettings {